casting for mac os

+++ src/VecSim/spaces/IP/IP_SVE_FP64.h
@@ -54,7 +54,8 @@ double FP64_InnerProductSIMD_SVE(const void *pVect1v, const void *pVect2v, size_
     }
 
     if constexpr (partial_chunk) {
-        svbool_t pg = svwhilelt_b64(offset, dimension);
+        svbool_t pg =
+            svwhilelt_b64(static_cast<uint64_t>(offset), static_cast<uint64_t>(dimension));
         svfloat64_t v1 = svld1_f64(pg, pVect1 + offset);
         svfloat64_t v2 = svld1_f64(pg, pVect2 + offset);
         sum3 = svmla_f64_m(pg, sum3, v1, v2);
